+++ server/context_key.go
@@ -1,4 +1,4 @@
-package serverutils
+package server
 
 type ContextKey string
